Move `workspaces-preview` folder
