+++ client/web/src/enterprise/batches/batch-spec/edit/workspaces-preview/PreviewPromptIcon.tsx
