+++ client/web/src/enterprise/batches/batch-spec/edit/workspaces-preview/WorkspacesPreviewListItem.story.tsx
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/react'
 import { noop } from 'lodash'
 
-import { WebStory } from '../../../../components/WebStory'
+import { WebStory } from '../../../../../components/WebStory'
 
 import { mockWorkspace } from './WorkspacesPreview.mock'
 import { WorkspacesPreviewListItem } from './WorkspacesPreviewListItem'
